rename: 피들로를 참고하여 데이터베이스 관련 함수 파일명을 변경한다
